修复专题页bug，数据逻辑带调整

+++ app/src/main/java/com/qiang/qiangguide/custom/channel/ChannelItem.java
@@ -34,14 +34,14 @@ public class ChannelItem implements Serializable {
     }
 
     public ChannelItem(int id, String name, int orderId,int selected) {
-        this.id = Integer.valueOf(id);
+        this.id = id;
         this.name = name;
-        this.orderId = Integer.valueOf(orderId);
-        this.selected = Integer.valueOf(selected);
+        this.orderId = orderId;
+        this.selected = selected;
     }
 
     public int getId() {
-        return this.id.intValue();
+        return this.id;
     }
 
     public String getName() {
@@ -49,7 +49,7 @@ public class ChannelItem implements Serializable {
     }
 
     public int getOrderId() {
-        return this.orderId.intValue();
+        return this.orderId;
     }
 
     public Integer getSelected() {
@@ -57,7 +57,7 @@ public class ChannelItem implements Serializable {
     }
 
     public void setId(int paramInt) {
-        this.id = Integer.valueOf(paramInt);
+        this.id = paramInt;
     }
 
     public void setName(String paramString) {
@@ -65,7 +65,7 @@ public class ChannelItem implements Serializable {
     }
 
     public void setOrderId(int paramInt) {
-        this.orderId = Integer.valueOf(paramInt);
+        this.orderId = paramInt;
     }
 
     public void setSelected(Integer paramInteger) {
